# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  notifications.NotificationActor:
    properties:
      avatar_id:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  notifications.NotificationItem:
    properties:
      actor:
        $ref: '#/definitions/notifications.NotificationActor'
      body:
        type: string
      created_at:
        type: string
      entity_id:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      read_at:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  notifications.NotificationListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/notifications.NotificationItem'
        type: array
      unreadCount:
        type: integer
    type: object
...
host: localhost:5000
info:
  contact: {}
  description: API documentation for the Chen, a music compatibility and social platform.
  title: Chen API
  version: "1.0"
paths:
...
  /notifications:
    get:
      description: Fetch all notifications for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/notifications.NotificationListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: List Notifications
      tags:
      - notifications
  /notifications/{id}/read:
    post:
      description: Mark a specific notification as read
      parameters:
      - description: Notification ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Mark Notification as Read
      tags:
      - notifications
  /notifications/unread-count:
    get:
      description: Get the number of unread notifications for the current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Get Unread Notification Count
      tags:
      - notifications
...
swagger: "2.0"